requirements.txt
geojson
geojsonio
requests
PyYAML